Add pulsing spin animation
The Font Awesome CSS framework has a pulse animation that rotates an
icon in 8 steps, which matches the structure of it's pulsing spinner.
This has been implemented with the "pulse" token here. Since Font
Awesome seems to have the only dotted pulsing spinner out of the current
font set, this seems a reasonable generic implementation at the moment.
If at some point spinners with different pulses need to be handled, then
a pulse property can be introduced for the icons, with 8 as the default
count.