Add IconImageView and IconImageButton
Although single icons can be rendered as text in text views, it requires
setting a fixed size. Using a drawable allows it to automatically scale
according to the size of the widget. Therefore, this commit introduces
IconImageView and IconImageButton to enable using the drawable in XML.
They also define custom attributes for the icon properties, which
improves readability and extendibility. Furthermore, extending ImageView
and ImageButton provides easy inoperability with other drawables (i.e
when switching between XML drawable resources and IconDrawable).